+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Added method `SearchUsers`
 - Added method `SearchChats`
 - Added method `SearchMessages`
-- Added fields `RootChatID`, `ChangedAt`, `DeletedAt` and  to `Message` struct
+- Added fields `RootChatID`, `ChangedAt`, and `DeletedAt` to `Message` struct
 
 ### [0.22.1](https://kaos.sh/pachca/0.22.1)